[MNT] deduplicate _check_param_grid into sktime.utils.sklearn
