+++ sktime/forecasting/model_selection/_gridsearch.py
@@ -1,12 +1,11 @@
 # copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
 """Grid search forecaster."""
 
-from collections.abc import Sequence
-
 import numpy as np
 from sklearn.model_selection import ParameterGrid
 
 from sktime.forecasting.model_selection._base import BaseGridSearch
+from sktime.utils.sklearn._model_selection import _check_param_grid
 
 
 class ForecastingGridSearchCV(BaseGridSearch):
@@ -311,33 +310,9 @@ class ForecastingGridSearchCV(BaseGridSearch):
         )
         self.param_grid = param_grid
 
-    def _check_param_grid(self, param_grid):
-        """_check_param_grid from sklearn 1.0.2, before it was removed."""
-        if hasattr(param_grid, "items"):
-            param_grid = [param_grid]
-
-        for p in param_grid:
-            for name, v in p.items():
-                if isinstance(v, np.ndarray) and v.ndim > 1:
-                    raise ValueError("Parameter array should be one-dimensional.")
-
-                if isinstance(v, str) or not isinstance(v, (np.ndarray, Sequence)):
-                    raise ValueError(
-                        f"Parameter grid for parameter ({name}) needs to"
-                        f" be a list or numpy array, but got ({type(v)})."
-                        " Single values need to be wrapped in a list"
-                        " with one element."
-                    )
-
-                if len(v) == 0:
-                    raise ValueError(
-                        f"Parameter values for parameter ({name}) need "
-                        "to be a non-empty sequence."
-                    )
-
     def _run_search(self, evaluate_candidates):
         """Search all candidates in param_grid."""
-        self._check_param_grid(self.param_grid)
+        _check_param_grid(self.param_grid)
         return evaluate_candidates(ParameterGrid(self.param_grid))
 
     @classmethod
